feat: musics collection, docker-compose

+++ src/models/musica_model.py
@@ -8,12 +8,20 @@ def criar_musica(musica):
 
 
 def ler_musica(id):
-    return musicas.find_one({"ID": id})
+    musica_encontrada = musicas.find_one({"ID": id})
+    if musica_encontrada:
+        return musica_encontrada
+
+
+def ler_musicas():
+    return list(musicas.find({}))
 
 
 def atualizar_musica(id, novos_valores):
-    musicas.update_one({"ID": id}, {"$set": novos_valores})
+    if ler_musica(id):
+        musicas.update_one({"ID": id}, {"$set": novos_valores})
 
 
 def deletar_musica(id):
-    musicas.delete_one({"ID": id})
+    if ler_musica(id):
+        musicas.delete_one({"ID": id})
